requirements.txt
arrow
backports.lzma
python-ceilometerclient
python-cinderclient
python-keystoneclient
python-novaclient
python-swiftclient
requests
swift>=2.0.0
tabulate
